requirements.txt
Flask==1.0.2
gunicorn==19.6.0
numpy==1.15.4
Pillow==5.3.0
https://download.pytorch.org/whl/cpu/torch-1.0.0-cp37-cp37m-linux_x86_64.whl
torchvision==0.2.1
